HW#5 v.1.2 1. Время показывает корректно 2. Фон не смещается при появлении клавиатуры

+++ app/src/main/java/ru/cocovella/WeatherApp/Model/ForecastModel/List.java
@@ -7,7 +7,7 @@ public class List {
 
     @SerializedName("dt")
     @Expose
-    private Integer dt;
+    private long dt;
     @SerializedName("main")
     @Expose
     private Main main;
@@ -30,11 +30,11 @@ public class List {
     @Expose
     private String dtTxt;
 
-    public Integer getDt() {
+    public long getDt() {
         return dt;
     }
 
-    public void setDt(Integer dt) {
+    public void setDt(long dt) {
         this.dt = dt;
     }
